Useful method extracted to a better place.
